Query::Archive: move obsid check for istoday here
Move the check for "obsid" being defined for "istoday" from
OMP::DB::Archive to the "istoday" method (in OMP::Query::Archive).
This should make this method safer to call in general because typical
obsid-based queries will not have dates, so "istoday" would trigger
an error by calling the "daterange" method.