DB::Archive::Cache: check for simple q. on store
This module already checks that the query is "simple" before retrieving
cached information (in "retrieve_archive") so that we don't get results
which don't match the query.  Update the "store_archive" method so that
it too checks the query is "simple" before writing a cache file.  This
is to avoid a cache file for a "non-simple" query being written such
that subsequent retrievals only get a subset of the expected results,
i.e. those matching the first query.