DB::MSB: directly check if first-accept msg. sent
When an MSB is accepted, and trying to determine whether the
"first accept" message needs to be sent, query the feedback
database instead of the MSB done table.  In this was we
can directly check whether the project has been sent the
message.  Otherwise multiple messages may be received
if one of the accepts failed such that it does not appear
in the MSB done table.