CGIComponent::ITCLink: add module to gen. ITC URLs
This adds a new module with a method to create links to the
integration time calculators hosted by the Hedwig proposal
system.  The script "misc/maint/get_itc_params.pl" can be
used to extract information about the calculators from
Hedwig, helping to ensure that we generate lists of the
correct parameters in the correct order.  For some parameters
(e.g. number of grid points, continuum mode) the necessary
information is not included in OMP::Info::Obs objects,
so not all parameters are filled correctly.