Translator::Headers::JCMT: conf. forcing JCMTCAL
This makes the forcing of the header PROJECT=JCMTCAL configurable
for "non-science" observations, with an initial configuration
value of 0 (i.e. no longer force these observations to be JCMTCAL).
The observations in question are those which are: marked as
standards, not of type "science" or "raw", or use "autoTarget".
The reason is to allow these kinds of observations to be performed
in E&C projects and not have them marked as JCMTCAL.