CGIComponent::NightRep: incl. skipped shifts in NR
When storing time accounting information, construct
OMP::Project::TimeAcct objects for all shifts, including those
marked to skip.  These are kept separate until after the
database has been updated, and are then included in the
group added to the night report object.  This is so
that, if the night report is being emailed, it contains
information for all shifts.