Query::Fault: add matching_responses_only option
This adds the option to retrieve only fault responses which
match the query parameters, rather than complete faults
with all responses included.

Also update the documentation for the "sql" method -- it seemed
to refer to the result only including the matching responses
whereas the method actually performs the match in a subquery
to find relevant faults and then retrieves everything.