NightRep: include fallback to fault 'date' if loss
Previously OMP::NightRep->faults queried for faults occurring
during the reporting period, falling back to the date of filing
for the original fault filing only.  Now also apply this fallback
to all responses with time loss.  This is so that the report
includes additional losses if the time of occurrence was not
entered.