Info::Obs, Info::Obs::TimeGap: use combined @DATA
Combine the information about each status (name and order and CSS class)
into a single data structure.  For now generate the previous listings
from this structure.  This is intended to make the information easier
to maintain and extend since all of the parameters for a given status
are kept together.