client/obslog: Use Info obj. obs/gap status lists
Instead of individually creating all of the observation and time
gap status radio buttons, generate these in a loop using lists
of status options from the OMP::Info::Obs and OMP::Info::Obs::TimeGap
classes.  A new method "get_status_options" was added to both
of these classes so that client/obslog.pl can call it on either
type of entry (observation or gap) and get the correct list of
statuses to show.