CGIPage::Fault: allow image upload to responses
The images are simply written into a response-numbered subdirectory
of the existing fault image directory.  A scaled version is written
if the size exceeds configured maximum dimensions.
OMP::Fault::Response gained an "images" attribute.  This is currently
used to store an array of hashes but a specific object type could
be added for this if needed.  Currently the list of images
is just read from the file system.  If this is too slow then
a database table for fault images could be added.